--- Final_Project/Test.py
@@ -1,9 +1,0 @@
-UD = {"hellrungj":["Natioh22",False]}
-print(UD)
-print(UD.items()[0])
-UD["praters"] = ["zackfair",False]
-print(UD)
-print(UD.items()[0])
-print(UD.items()[1])
-print(UD.items()[1][0])
-print(UD.items()[1][1][0])
