+++ Final_Project/app/allImports.py
@@ -1,0 +1,7 @@
+from __future__ import print_function
+
+#Include all imports in this file; it will be called at the beginning of all files.
+
+from app import models
+from models import *                # all the database models
+from app.logic.switch import switch # implements switch/case statements
