+++ Final_Project/app/logic/__init__.py
@@ -1,0 +1,1 @@
+from app.logic.switch import switch
